Fixed Doppelganger style level reset bug for Vergil and probably Dante too.
This was caused by UpdateStyle being called twice and this in turn was caused by a missing magical controller variable which was not set in Arcade mode.

Updated Arcade variables.

Updated game vars. Removed MODE_HEAVEN_OR_HELL as it does not really exist.

+++ Mary/Config.h
@@ -10,8 +10,8 @@ struct CONFIG
 		struct
 		{
 			bool   enable         = false;
-			uint8  mission        = 1;
-			uint8  mode           = MODE_NORMAL;
+			uint32 mission        = 1;
+			uint32 mode           = MODE_NORMAL;
 			uint16 room           = 0;
 			bool   ignoreRoom     = false;
 			uint16 position       = 0;
